Spawn players in PlayerSpawn position per Level.
Note: There is some indirection required due to how we load an entire
World's LevelSet rather than individual LevelSelection, and seems
bevy_ecs_ldtk may not automatically attach LdtkEntity Components to
Handle<LdtkLevel> for us.

Note: We use some more indirection using states to load entities as it
seems the first iteration of the loop isn't able to actually find the
LdtkEntity Components.  Unsure if this is a bug with bevy_ecs_ldtk or as
per design